# shared/openapi/openapi.yaml
openapi: 3.0.2
paths:
...
  /me/dxnet-jobs/active:
    get:
      tags: *a5
      parameters:
        - name: authorization
          in: header
          required: true
          schema:
            type: string
      operationId: job.getActiveByFriendCode
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                type: object
                properties:
                  job:
                    type: object
                    properties:
                      id:
                        type: string
                      friendCode:
                        type: string
                      jobType:
                        type: string
                        enum:
                          - send_friend_request
                          - accept_friend_request
                          - update_score
                          - get_user_recent_event
                      priority:
                        type: integer
                      skipUpdateScore:
                        type: boolean
                      botUserFriendCode:
                        type: string
                        nullable: true
                      friendRequestSentAt:
                        type: string
                        nullable: true
                      friendRequestWaitStartedAt:
                        type: string
                        nullable: true
                      status:
                        type: string
                        enum:
                          - queued
                          - processing
                          - completed
                          - failed
                          - canceled
                      stage:
                        type: string
                        enum:
                          - send_request
                          - wait_acceptance
                          - wait_user_request
                          - accept_request
                          - update_score
                          - get_user_recent_event
                      profile:
                        nullable: true
                      error:
                        type: string
                        nullable: true
                      executing:
                        type: boolean
                      scoreProgress:
                        type: object
                        properties:
                          completedDiffs:
                            type: array
                            items:
                              type: integer
                              minimum: 0
                              maximum: 14
                          totalDiffs:
                            type: integer
                            minimum: 0
                        required:
                          - completedDiffs
                          - totalDiffs
                        nullable: true
                      updateScoreDuration:
                        type: number
                        nullable: true
                      autoExportResult:
                        type: object
                        properties:
                          divingFish:
                            type: object
                            properties:
                              status:
                                type: string
                              message:
                                type: string
                            required:
                              - status
                            nullable: true
                          lxns:
                            type: object
                            properties:
                              status:
                                type: string
                              message:
                                type: string
                            required:
                              - status
                            nullable: true
                        nullable: true
                      isAuthenticated:
                        type: boolean
                      cabinetScoreMap:
                        type: object
                        additionalProperties:
                          type: object
                          properties:
                            achievement:
                              type: number
                            dxScore:
                              type: number
                          required:
                            - achievement
                            - dxScore
                        nullable: true
                      diffsToScrape:
                        type: array
                        items:
                          type: integer
                        nullable: true
                      runAt:
                        type: string
                        nullable: true
                      createdAt:
                        type: string
                      updatedAt:
                        type: string
                    required:
                      - id
                      - friendCode
                      - jobType
                      - skipUpdateScore
                      - status
                      - stage
                      - createdAt
                      - updatedAt
                    nullable: true
                required:
                  - job
...
  /workers/dxnet/jobs/{jobId}/api-logs:
    post:
      tags: *a5
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      operationId: job.addApiLogs
      requestBody:
        description: Body
        content:
          application/json:
            schema:
              type: object
              properties:
                logs:
                  type: array
                  items:
                    type: object
                    properties:
                      url:
                        type: string
                      method:
                        type: string
                      statusCode:
                        type: number
                      bodySize:
                        type: integer
                        minimum: 0
                        nullable: true
                    required:
                      - url
                      - method
                      - statusCode
              required:
                - logs
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
...
  /admin/dashboard/job-trend:
    get:
      tags: *a7
      parameters:
        - name: hours
          in: query
          schema:
            type: string
      operationId: admin.getJobTrend
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                nullable: true
...
  /admin/dxnet-jobs:
    get:
      tags: *a7
      parameters:
        - name: friendCode
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            default: 1
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          schema:
            default: 10
            type: integer
            minimum: 1
            maximum: 100
      operationId: admin.searchJobs
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                nullable: true
...
  /admin/bots/{friendCode}/remark:
    patch:
      tags: *a7
      parameters:
        - name: friendCode
          in: path
          required: true
          schema:
            type: string
      operationId: admin.updateBotRemark
      requestBody:
        description: Body
        content:
          application/json:
            schema:
              type: object
              properties:
                remark:
                  type: string
                  nullable: true
              required:
                - remark
      responses:
        "200":
          description: "200"
  /admin/bots/{friendCode}/cabinet-user-id:
    patch:
      tags: *a7
      parameters:
        - name: friendCode
          in: path
          required: true
          schema:
            type: string
      operationId: admin.updateBotCabinetUserId
      requestBody:
        description: Body
        content:
          application/json:
            schema:
              type: object
              properties:
                cabinetUserId:
                  type: integer
                  minimum: 0
                  exclusiveMinimum: 0
                  nullable: true
              required:
                - cabinetUserId
      responses:
        "200":
          description: "200"
  /admin/bots/{friendCode}:
    delete:
      tags: *a7
      parameters:
        - name: friendCode
          in: path
          required: true
          schema:
            type: string
      operationId: admin.removeBot
      responses:
        "200":
          description: "200"
  /admin/dxnet-jobs/{jobId}/api-logs:
    get:
      tags: *a7
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      operationId: admin.getJobApiLogs
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                type: array
                items:
                  nullable: true
...
info:
  title: Maimai Score Hub API
  version: 1.0.0
  description: Generated from @maimai-score-hub/shared ts-rest contracts
servers:
  - url: /api/v1